Move .h files to .hpp.
