+++ aproj/aproj-init.cpp
@@ -6,8 +6,8 @@
 #include <sstream>
 #include <vector>
 
-#include <antler/project/project.h>
-#include <aproj-common.h>
+#include <antler/project/project.hpp>
+#include <aproj-common.hpp>
 
 
 /// Print usage information to std::cout and return 0 or, optionally - if err is not-empty() - print to std::cerr and return -1.
