+++ common/antler/string/trim.hpp
