+++ common/src/so_support.hpp
