+++ project/antler/project/project.hpp
@@ -3,7 +3,7 @@
 
 /// @copyright See `LICENSE` in the root directory of this project.
 
-#include <antler/project/object.h>
+#include <antler/project/object.hpp>
 #include <filesystem>
 #include <optional>
