+++ project/src/cmake.hpp
@@ -3,7 +3,7 @@
 
 /// @copyright See `LICENSE` in the root directory of this project.
 
-#include <antler/project/semver.h>
+#include <antler/project/semver.hpp>
 
 #include <filesystem> // path
