+++ project/src/key.cpp
@@ -1,6 +1,6 @@
 /// @copyright See `LICENSE` in the root directory of this project.
 
-#include <key.h>
+#include <key.hpp>
 #include <ostream>
 
 /// Macro to automate insertion and extraction of key::word enum.
