+++ project/src/langague.cpp
@@ -1,6 +1,6 @@
 /// @copyright See `LICENSE` in the root directory of this project.
 
-#include <antler/project/language.h>
+#include <antler/project/language.hpp>
 
 // Mapping of enum to string, please maintain string as lowercase.
 #define LANGUAGE_CASE_OF                        \
