+++ project/src/object.cpp
@@ -1,6 +1,6 @@
 /// @copyright See `LICENSE` in the root directory of this project.
 
-#include <antler/project/object.h>
+#include <antler/project/object.hpp>
 
 #include <algorithm> // find_if()
